[CORE] SingUp, LogIn, LogOut implemented. Remove and Add credit card implemented. Modify user implemented

+++ Hamp/Hamp/Factories/LaundryServicesFactory.swift
@@ -41,7 +41,7 @@ enum ServiceType: String {
 }
 
 struct LaundryServicesFactory {
-    static func service(by type: ServiceType, price: Int = 1) -> Service {
+    static func service(by type: ServiceType, price: Float = 1) -> Service {
         return LaundryService.init(identifier: type.rawValue, price: price, name: type.name)
     }
 }
